+++ src/extraConsoleCommands.ts
@@ -1,4 +1,3 @@
-import { setDebugBabyType } from "./classes/features/babySelection/v";
 import { debugFunction, setDebugBaby } from "./debugFunction";
 import { mod } from "./mod";
 
@@ -6,7 +5,6 @@ export function enableExtraConsoleCommandsBabiesMod(): void {
   mod.addConsoleCommand("baby", baby);
   mod.addConsoleCommand("baby2", baby2);
   mod.addConsoleCommand("babyDebug", babyDebug);
-  mod.addConsoleCommand("babyRandom", babyRandom);
 }
 
 /** "baby #" will restart as the specified baby. */
@@ -21,9 +19,4 @@ function baby2(params: string) {
 
 function babyDebug() {
   debugFunction();
-}
-
-function babyRandom() {
-  setDebugBabyType(undefined);
-  Isaac.ExecuteCommand("restart");
 }